Adjust scaling of pie charts when zooming out
When the user zooms out, the pie charts as a whole will now scale
according to the exponential function e^(-resolutionOfMap).
For this to work, caching of pie charts had to be disabled starting from
a given number of remaining people (default: 500). A new config-table
entry was created in the database.